Check if product app is unbundled from property.
Some of devices launched with old version of android (and maybe not
treblized) can come with no /product/lib. However in current we are
checking if product app is unbundled by checking /product/lib from
search path, this can lead to wrong check of unbundled. To fix this, I
created new property (ro.product.apps.unbundled) to decide if product
app is unbundled.

Bug: 129011845
Test: m -j && tested from crosshatch
Change-Id: I6880119b75e1abb6cbe6f30caed1cac345bd1048