Make ota.warm_set a sysprop API
The property is set to inform kernel to do a warm_reset on the next
reboot. This is useful to persist the logs to debug device boot
failures. More details in http://go/rvc-ota-persist-logs.

Since the property is set in core domain and read by vendor init, we
want to make it stable.

Bug: 143489994
Test: Build

Change-Id: I43cfbe4a6af11ed470dd890f23e6ef9c8e92b7f6