Add sysprop to enable/disable Global drag and drop
DD: go/global_drag_and_drop

Using this sysprop to use as an internal flag to flag guard
implementation for global drag and drop.
Will use this flag to guard implementation in DragDropController
and InputDispatcher to enable the global drag feature.

Bug: 220109830

Test: manually by setting the property using cmd
'adb shell setprop persist.input.globaldraganddrop <value>' and
testing if the value is refected in the code when accessed.

Change-Id: I8ef16fc4d5a89669fa0d03c5633191793e153f51